Add video decoder and rendering layer for Android
This uses the MediaCodec API to decode incoming h264 NAL units, which
renders to a SurfaceTexture. We then take that SurfaceTexture and
use it with QSGRenderNode to render inside the Qt scene graph.

This has some odd quirks at the moment, in the past it has been fairly
smooth and low latency but is showing a little of both right now.